requirements.txt
Django==5.0.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
pandas==2.1.3
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.1.0
plotly==5.18.0
python-dotenv==1.0.0
sentence-transformers
transformers
huggingface-hub
faiss-cpu
torch

